exercise 4.21

+++ bloglist/controllers/blogs.js
@@ -41,10 +41,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   const blog = await Blog.findById(request.params.id)
   if (!blog) {
     return response.status(404).end()
   }
+  console.log(blog.user.toString())
+  if (decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
